requirements.txt
schedule==1.2.1
yagmail==0.15.293
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cryptography==42.0.5
pyOpenSSL==24.0.0
ratelimit==2.2.1
validators==0.22.0
python-json-logger==2.0.7
pycryptodome==3.20.0
lxml==5.1.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
gspread==6.0.2
oauth2client==4.1.3 